Modified organization. Started GUI.
